improved

+++ fk_conn.c
@@ -235,7 +235,7 @@ int fk_conn_req_parse(fk_conn *conn)
 
 		if (conn->idx_flag == 1) {
 			start = fk_buf_payload_start(rbuf);
-			len = (int)fk_conn_arglen_get(conn, conn->arg_idx);
+			len = (intptr_t)fk_conn_arglen_get(conn, conn->arg_idx);
 			fk_log_debug("saved arg_len: %lu\n", len);
 
 			if (fk_buf_payload_len(rbuf) >= len + 2) {
